requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
pydantic>=2